requirements.txt
asgiref==3.7.2
autopep8==1.6.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==0.5.0
Django==4.2.11
django-environ==0.8.1
google-api-core==2.17.1
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gunicorn==20.1.0
idna==3.6
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pytz==2021.3
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==5.3.0
